refactor: split fielding-view into folder, imperative editor handles
- Break 865-line monolith into 5 files under fielding-view/
- Replace onChange callback with useImperativeHandle + getCurrentData()
- Dialog pulls editor data imperatively on save (no unmount timing issues)
- Remove syncEnabledRef and setTimeout hacks from edit-views-dialog
- Add README documenting how to implement a new session view
