+++ components/softball/session-views/fielding-view/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./fielding-viewer";
+export { SoftballFieldingEditor } from "./fielding-editor";
