+++ components/sports/session-views/interfaces.ts
@@ -1,7 +1,7 @@
-import type { ComponentType } from "react";
+import type { ComponentType, Ref } from "react";
 import type { SignupRow } from "@/components/sports/session-signups-table";
 
-/** Props passed to every session view component. */
+/** Props passed to every session view component (read-only viewer). */
 export interface SessionViewProps {
     signups: SignupRow[];
     teamMemberIds: Set<string>;
@@ -11,14 +11,26 @@ export interface SessionViewProps {
     viewData: unknown;
 }
 
-/** Props passed to every session view editor component. */
+/**
+ * Imperative handle exposed by every editor via useImperativeHandle.
+ * The dialog calls getCurrentData() to pull the editor's state on save.
+ */
+export interface SessionViewEditorHandle {
+    getCurrentData: () => unknown;
+}
+
+/**
+ * Props passed to every session view editor component.
+ * Editors own their state internally and expose it via the ref handle.
+ * See fielding-view/index.ts for full documentation on implementing a new view.
+ */
 export interface SessionViewEditorProps {
     signups: SignupRow[];
     teamMemberIds: Set<string>;
-    /** Current saved data for this view. */
+    /** Initial data for this view (from the DB). Do not mutate — use as seed for local state. */
     viewData: unknown;
-    /** Called when the editor's data changes. Parent batches the save. */
-    onChange: (data: unknown) => void;
+    /** Ref for the dialog to pull data imperatively via getCurrentData(). */
+    ref?: Ref<SessionViewEditorHandle>;
 }
 
 /**
